+++ metrics/retry.go
@@ -4,7 +4,6 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
-// Retry makes table.RetryTrace with New publishing
-func Retry(config Config) (t trace.Retry) {
+func retry(config Config) (t trace.Retry) {
 	return t
 }