+++ tests/slo/xorm/storage.go
@@ -140,7 +140,7 @@ func (s *Storage) Read(ctx context.Context, id generator.RowID) (row generator.R
 		},
 		retry.WithIdempotent(true),
 		retry.WithTrace(
-			trace.Retry{
+			&trace.Retry{
 				OnRetry: func(info trace.RetryLoopStartInfo) func(trace.RetryLoopIntermediateInfo) func(trace.RetryLoopDoneInfo) {
 					return func(info trace.RetryLoopIntermediateInfo) func(trace.RetryLoopDoneInfo) {
 						return func(info trace.RetryLoopDoneInfo) {
@@ -174,7 +174,7 @@ func (s *Storage) Write(ctx context.Context, row generator.Row) (attempts int, e
 		},
 		retry.WithIdempotent(true),
 		retry.WithTrace(
-			trace.Retry{
+			&trace.Retry{
 				OnRetry: func(info trace.RetryLoopStartInfo) func(trace.RetryLoopIntermediateInfo) func(trace.RetryLoopDoneInfo) {
 					return func(info trace.RetryLoopIntermediateInfo) func(trace.RetryLoopDoneInfo) {
 						return func(info trace.RetryLoopDoneInfo) {